simple websocket viewer added.

+++ plugins/server/external/simple-web-crypto/base64.h
@@ -17,7 +17,7 @@ class Base64 {
 
     out->resize(EncodedLength(in));
 
-    int input_len = in.size();
+    size_t input_len = in.size();
     std::string::const_iterator input = in.begin();
 
     while (input_len--) {
@@ -100,7 +100,7 @@ class Base64 {
     unsigned char a3[3];
     unsigned char a4[4];
 
-    int input_len = in.size();
+	size_t input_len = in.size();
     std::string::const_iterator input = in.begin();
 
     out->resize(DecodedLength(in));
@@ -195,7 +195,7 @@ class Base64 {
     return (out == (out_begin + decoded_length));
   }
 
-  static int DecodedLength(const char *in, size_t in_length) {
+  static size_t DecodedLength(const char *in, size_t in_length) {
     int numEq = 0;
 
     const char *in_end = in + in_length;
@@ -204,9 +204,9 @@ class Base64 {
     return ((6 * in_length) / 8) - numEq;
   }
 
-  static int DecodedLength(const std::string &in) {
+  static size_t DecodedLength(const std::string &in) {
     int numEq = 0;
-    int n = in.size();
+	size_t n = in.size();
 
     for (std::string::const_reverse_iterator it = in.rbegin(); *it == '='; ++it) {
       ++numEq;
@@ -215,11 +215,11 @@ class Base64 {
     return ((6 * n) / 8) - numEq;
   }
 
-  inline static int EncodedLength(size_t length) {
+  inline static size_t EncodedLength(size_t length) {
     return (length + 2 - ((length + 2) % 3)) / 3 * 4;
   }
 
-  inline static int EncodedLength(const std::string &in) {
+  inline static size_t EncodedLength(const std::string &in) {
     return EncodedLength(in.length());
   }
